fixes search
creates a separate author column for search that is indexed but not visible. this allows for search to work even when the author name in the metadata appears as a link. also fixes the issue of not counting links correctly in the subject cloud by replacing the fields in the subject cloud with subject;searchau. that is in the theme.yml file